refactor : pr#42 리뷰 반영
1. 안쓰는 코드 주석 처리된 것 삭제.
2. 사용하지 않는 어노테이션 의존성 삭제
3. 패키지 이름을 snake_case로 작성

<다음 PR에 해결할 일>
1. create_table.sql 내 스키마명 상수화
2. 통합 테스트 구현 사항 재고려
