+++ docker-compose.yml
@@ -2,15 +2,15 @@ version: '3.8'
 # convention : container_name/depends_on/image or build/restart/ports/environment
 services:
   rediscache:
-    build: src/test/resources/testcontainers/cache
+    build: cache
     restart: always
 
   redissession:
-    build: src/test/resources/testcontainers/session
+    build: session
     restart: always
 
   database:
-    build: src/test/resources/testcontainers/database
+    build: database
     restart: always
     environment:
       - MYSQL_ROOT_PASSWORD=root
@@ -35,3 +35,4 @@ services:
       - MYSQL_ROOT_PASSWORD=root
       - MYSQL_USER=resell_platform
       - MYSQL_PASSWORD=resell1234
+      - DB_SCHEMA=v_1
