+++ src/main/java/flab/resellPlatform/RedisSessionConfig.java
@@ -1,20 +1,13 @@
 package flab.resellPlatform;
 
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Lazy;
-import org.springframework.context.annotation.Primary;
-import org.springframework.data.redis.cache.RedisCacheConfiguration;
-import org.springframework.data.redis.cache.RedisCacheManager;
 import org.springframework.data.redis.connection.RedisConnectionFactory;
 import org.springframework.data.redis.connection.RedisStandaloneConfiguration;
 import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory;
 import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.data.redis.serializer.GenericJackson2JsonRedisSerializer;
-import org.springframework.data.redis.serializer.RedisSerializationContext;
 import org.springframework.data.redis.serializer.StringRedisSerializer;
 
 import java.time.Duration;
@@ -30,16 +23,12 @@ public class RedisSessionConfig {
     @Value("${spring.session.redis.master.port}")
     int masterPort;
 
-//    @Value("${spring.session.redis.master.password}")
-//    String masterPW;
-
     // 세션
     @Bean
     public RedisConnectionFactory redisSessionConnectionFactory() {
         RedisStandaloneConfiguration configuration = new RedisStandaloneConfiguration();
         configuration.setHostName(masterHostName);
         configuration.setPort(masterPort);
-//        configuration.setPassword(masterPW);
         return new LettuceConnectionFactory(configuration);
     }
