+++ src/test/java/flab/integration_test/IntegrationTest.java
@@ -1,4 +1,4 @@
-package flab.IntegrationTest;
+package flab.integration_test;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
